Split search tests into separate files for EO3 and legacy tests.
